abc_bank_py/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "abc-bank-py"
version = "0.1.0"
description = "Python port of the abc_bank Spring Boot demo"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.110",
    "email-validator>=2.0",
    "uvicorn[standard]>=0.27",
    "sqlalchemy>=2.0",
    "pydantic>=2.6",
    "pydantic-settings>=2.2",
    "pyjwt>=2.8",
    "bcrypt>=4.0",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
dev = [
    "pytest>=8",
    "httpx>=0.27",
]

[tool.setuptools.packages.find]
include = ["app*"]
